# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Almoxarifado Inteligente — API
  version: 1.0.0
  description: |
    API REST do sistema de gestão de almoxarifado industrial. Autenticação por
    Bearer token (256-bit) com RBAC (ADMIN_MASTER, OPERADOR, COMPRAS, CONSULTA),
    rate limiting por IP, auditoria automática de mutações sensíveis e backups
    locais + Google Drive.

    **Autenticação:** faça `POST /api/login` e envie `Authorization: Bearer <token>`.
    Rota pública de health: `GET /healthz`.
  license:
    name: MIT
    url: https://github.com/VictorHugoEng/AlmoxarifadoProject/blob/main/LICENSE
  contact:
    name: Victor Hugo Eng
    url: https://github.com/VictorHugoEng
servers:
  - url: http://localhost:3000
    description: Desenvolvimento local
  - url: https://almoxarifado.onrender.com
    description: Produção (Render)
...
paths:
...
  /api/nuvem/oauth2/callback:
    get:
      tags: [Nuvem e Backup]
      summary: Retorno do Google (troca code por token e envia 1º backup)
      security: []
      parameters:
        - { name: code, in: query, required: true, schema: { type: string } }
      responses:
        '302':
          description: Redireciona para /login.html?nuvem=ok (ou =erro)
  /api/nuvem/config:
    post:
      tags: [Nuvem e Backup]
      summary: Salva credenciais do app Google (ADMIN_MASTER)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_id, client_secret]
              properties:
                client_id: { type: string }
                client_secret: { type: string }
      responses:
        '200':
          description: Credenciais salvas
  /api/nuvem/enviar:
    post:
      tags: [Nuvem e Backup]
      summary: Envia backup para o Google Drive agora (ADMIN_MASTER)
      responses:
        '200':
          description: Backup enviado (ou nuvem já atualizada)
        '400':
          description: Sem credenciais ou sem tokens
...
  /api/atualizacao/aplicar:
    post:
      tags: [Nuvem e Backup]
      summary: Aplica pacote OTA baixado da nuvem e reinicia (ADMIN_MASTER)
      responses:
        '200':
          description: Atualização aplicada, reiniciando
        '404':
          description: Nenhum pacote de atualização na nuvem
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []